MelanoMAP code repository/requirements.txt
albumentations==1.3.0
segmentation-models_pytorch==0.3.3
matplotlib==3.8.3
wandb==0.13.9
pandas==2.0.1
scikit-image==0.20.0
seaborn==0.12.2
openslide-python==1.3.1
h5py==3.10.0
torchvision
boto3==1.34.141
streamlit==1.36.0
stqdm==0.0.5
lifelines==0.30.0
scikit-survival==0.24.0